Restructure sts-mock directory
