+++ sts-mock/src/functions/sts-mock/services/logging/README.md
