+++ sts-mock/src/functions/sts-mock/services/logging/commonRegisteredLogs.ts
