+++ sts-mock/src/functions/sts-mock/services/logging/logger.ts
