+++ sts-mock/src/functions/sts-mock/services/logging/tests/mockLoggingAdapter.ts
