+++ sts-mock/src/functions/sts-mock/services/logging/types.ts
