+++ sts-mock/src/functions/sts-mock/token/handlerDependencies.ts
