+++ sts-mock/src/functions/sts-mock/token/keyRetriever/tests/mocks.ts
@@ -1,6 +1,10 @@
-import { successResult, Result, errorResult } from "../../../utils/result";
+import {
+  successResult,
+  Result,
+  errorResult,
+} from "../../../../../utils/result";
 import { IKeyRetriever, SigningKey } from "../keyRetriever";
-import { getMockSigningKey } from "../../../testUtils/getMockSigningKey";
+import { getMockSigningKey } from "../../../../../testUtils/getMockSigningKey";
 
 export class MockKeyRetrieverSuccessResult implements IKeyRetriever {
   async getKey(): Promise<Result<SigningKey>> {
