+++ sts-mock/src/functions/sts-mock/token/registeredLogs.ts
