+++ sts-mock/src/functions/sts-mock/token/tokenEncrypter/tests/mocks.ts
@@ -1,5 +1,9 @@
 import { ITokenEncrypter, JWE } from "../tokenEncrypter";
-import { errorResult, Result, successResult } from "../../../utils/result";
+import {
+  errorResult,
+  Result,
+  successResult,
+} from "../../../../../utils/result";
 
 export class MockTokenEncrypterSuccessResult implements ITokenEncrypter {
   async encrypt(): Promise<Result<JWE>> {
