+++ sts-mock/src/functions/sts-mock/token/tokenEncrypter/tests/tokenEncrypter.test.ts
