+++ sts-mock/src/functions/sts-mock/token/tokenHandler.ts
