+++ sts-mock/src/functions/sts-mock/token/tokenSigner/tokenSigner.ts
@@ -1,4 +1,4 @@
-import { errorResult, Result, successResult } from "../../utils/result";
+import { errorResult, Result, successResult } from "../../../../utils/result";
 import { JWTPayload, SignJWT } from "jose";
 import { KeyObject } from "node:crypto";
