+++ sts-mock/src/functions/sts-mock/token/validateServiceTokenRequest/validateServiceTokenRequest.ts
@@ -1,4 +1,4 @@
-import { errorResult, Result, successResult } from "../../utils/result";
+import { errorResult, Result, successResult } from "../../../../utils/result";
 
 export type IValidateServiceTokenRequest = (
   requestBody: string | null,
