+++ sts-mock/src/testUtils/getMockSigningKey.ts
@@ -1,4 +1,4 @@
-import { SigningKey } from "../token/keyRetriever/keyRetriever";
+import { SigningKey } from "../functions/sts-mock/token/keyRetriever/keyRetriever";
 import { createPrivateKey } from "node:crypto";
 
 export function getMockSigningKey(): SigningKey {
