+++ sts-mock/template.yaml
@@ -218,7 +218,7 @@ Resources:
         Target: es2022
         Sourcemap: false
         EntryPoints:
-          - src/token/tokenHandler.ts
+          - src/functions/sts-mock/token/tokenHandler.ts
     Properties:
       Description: Mock STS token function for exchanging an access token for a service token
       AutoPublishAlias: live
